+++ lib/puppet/provider/nsswitch/solaris.rb
@@ -14,7 +14,9 @@
 # limitations under the License.
 #
 
-Puppet::Type.type(:nsswitch).provide(:nsswitch) do
+Puppet::Type.type(:nsswitch).provide(:nsswitch,
+  :parent =>
+  Puppet::Type.type(:svccfg).provider(:svccfg)) do
   desc "Provider for name service switch configuration data"
   confine :operatingsystem => [:solaris]
   defaultfor :osfamily => :solaris, :kernelrelease => ['5.11', '5.12']
@@ -23,35 +25,52 @@ Puppet::Type.type(:nsswitch).provide(:nsswitch) do
   class << self; attr_accessor :nsswitch_fmri end
   Nsswitch_fmri = "svc:/system/name-service/switch"
 
+  mk_resource_methods
+
   def self.instances
     props = {}
     svcprop("-p", "config", Nsswitch_fmri).each_line.collect do |line|
-      fullprop, value = line.split(" ", 2).values_at(0,2)
+      fullprop, value = line.split(" ", 3).values_at(0,2)
       prop = fullprop.split("/")[1].intern
-      props[prop] = value \
+      props[prop] = value.strip.delete('\\') \
         if Puppet::Type.type(:nsswitch).validproperties.include? prop
     end
 
     props[:name] = "current"
+    props[:ensure] = :present
     return Array new(props)
   end
 
-  Puppet::Type.type(:nsswitch).validproperties.each do |field|
-    define_method(field) do
-      begin
-        out = svcprop("-p", "config/" + field.to_s,
-                      Nsswitch_fmri).strip()
-        out = out.delete("\\")
-      rescue
-        # if the property isn't set, don't raise an error
-        nil
-      end
-    end
+  def self.prefetch(resources)
+    binding.pry
+    resources.keys.each { |key|
+      things.find { |prop|
+        # key.to_s in case name uses newvalues and is converted to symbol
+        prop.name == key.to_s
+      }.tap { |provider|
+        next if provider.nil?
+        resources[key].provider = provider
+      }
+    }
+  end
 
-    define_method(field.to_s + "=") do |should|
+  def exists?
+    @property_hash[:ensure] == :present
+  end
+
+  def create
+    fail "can only be modified not created"
+  end
+  def destroy
+    fail "can only be modified not destroyed"
+  end
+
+  Puppet::Type.type(:nsswitch).validproperties.each do |field|
+    define_method(field.to_s + "=") do |value|
       begin
-        svccfg("-s", Nsswitch_fmri, "setprop",
-               "config/" + field.to_s, "=", '"' + should + '"')
+        should = munge_value(value,:astring)
+        svccfg("-s", Nsswitch_fmri, "setprop", "config/#{field}=#{should}")
+        return value
       rescue => detail
         fail "value: #{should.inspect}\n#{detail}\n"
       end